# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * النبض المهني - Professional Pulse API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type StudentReadinessLevel = typeof StudentReadinessLevel[keyof typeof StudentReadinessLevel];


export const StudentReadinessLevel = {
  beginner: 'beginner',
  developing: 'developing',
  ready: 'ready',
  advanced: 'advanced',
} as const;

export interface Student {
  id: number;
  name: string;
  studentId: string;
  major: string;
  year: number;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  overallScore: number;
  readinessLevel?: StudentReadinessLevel;
  createdAt: string;
}

export interface StudentInput {
  name: string;
  studentId: string;
  major: string;
  year: number;
  bio?: string;
  avatarUrl?: string;
}

export interface StudentUpdate {
  name?: string;
  major?: string;
  year?: number;
  bio?: string;
  avatarUrl?: string;
}

export type SkillCategory = typeof SkillCategory[keyof typeof SkillCategory];


export const SkillCategory = {
  programming: 'programming',
  design: 'design',
  data: 'data',
  management: 'management',
  communication: 'communication',
  other: 'other',
} as const;

export type SkillDemandLevel = typeof SkillDemandLevel[keyof typeof SkillDemandLevel];


export const SkillDemandLevel = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface Skill {
  id: number;
  name: string;
  category: SkillCategory;
  demandLevel: SkillDemandLevel;
  marketWeight: number;
  /** @nullable */
  description?: string | null;
}

export interface StudentSkill {
  id: number;
  studentId: number;
  skillId: number;
  skillName: string;
  category: string;
  /** 1-5 proficiency level */
  level: number;
  verified: boolean;
  score: number;
  /** @nullable */
  verifiedAt?: string | null;
}

export interface StudentSkillInput {
  skillId: number;
  level: number;
}

export type SkillGapPriority = typeof SkillGapPriority[keyof typeof SkillGapPriority];


export const SkillGapPriority = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface SkillGap {
  skillName: string;
  currentLevel: number;
  requiredLevel: number;
  gap: number;
  priority: SkillGapPriority;
  recommendation?: string;
}

export interface SkillGapAnalysis {
  studentId: number;
  overallReadiness: number;
  gaps: SkillGap[];
  strengths: string[];
}

export interface InvestmentValue {
  studentId: number;
  hourlyRateMin: number;
  hourlyRateMax: number;
  hourlyRateEstimate: number;
  currency: string;
  rarityScore: number;
  demandScore: number;
  marketInsight?: string;
}

export type CareerPathDemandLevel = typeof CareerPathDemandLevel[keyof typeof CareerPathDemandLevel];


export const CareerPathDemandLevel = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface CareerPath {
  sector: string;
  matchScore: number;
  roles: string[];
  demandLevel: CareerPathDemandLevel;
  salaryRange?: string;
  description?: string;
}

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  pending: 'pending',
  analyzing: 'analyzing',
  verified: 'verified',
  rejected: 'rejected',
} as const;

export interface Project {
  id: number;
  studentId: number;
  studentName: string;
  title: string;
  description: string;
  technologies: string[];
  /** @nullable */
  githubUrl?: string | null;
  /** @nullable */
  demoUrl?: string | null;
  status: ProjectStatus;
  trustStamp: boolean;
  /** @nullable */
  aiScore?: number | null;
  /** @nullable */
  integrityNotes?: string | null;
  createdAt: string;
}

export interface ProjectInput {
  studentId: number;
  title: string;
  description: string;
  technologies: string[];
  githubUrl?: string;
  demoUrl?: string;
}

export interface ProjectVerifyInput {
  codeSnippet: string;
  projectDescription?: string;
}

export interface ProjectVerification {
  projectId: number;
  passed: boolean;
  aiScore: number;
  summary: string;
  originality: number;
  integrityNotes?: string;
}

export interface QuizQuestion {
  id: number;
  question: string;
  options: string[];
}

export interface ProjectQuiz {
  projectId: number;
  questions: QuizQuestion[];
}

export interface QuizAnswer {
  questionId: number;
  answer: string;
}

export interface QuizSubmission {
  answers: QuizAnswer[];
}

export interface QuizResult {
  projectId: number;
  passed: boolean;
  score: number;
  feedback: string;
  trustStampGranted: boolean;
}

export interface CertifiedSkill {
  name: string;
  level: number;
  verified: boolean;
  aiVerified: boolean;
}

export interface CertifiedProject {
  title: string;
  technologies: string[];
  trustStamp: boolean;
}

export interface Certificate {
  id: number;
  studentId: number;
  studentName: string;
  major: string;
  overallScore: number;
  readinessLevel?: string;
  qrCode: string;
  skills: CertifiedSkill[];
  projects: CertifiedProject[];
  issuedAt: string;
}

export interface DashboardStats {
  totalStudents: number;
  verifiedProjects: number;
  averageReadiness: number;
  certificatesIssued: number;
  topSector: string;
  studentsReadyForWork?: number;
  skillsTracked?: number;
}

export interface TopSkill {
  name: string;
  demandLevel: string;
  studentsCount: number;
  marketDemand: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  project_verified: 'project_verified',
  skill_added: 'skill_added',
  certificate_issued: 'certificate_issued',
  quiz_passed: 'quiz_passed',
  student_joined: 'student_joined',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  message: string;
  studentName: string;
  createdAt: string;
}