# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: النبض المهني - Professional Pulse API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Submit a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Project submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}/verify:
    post:
      operationId: verifyProject
      tags: [projects]
      summary: Run AI integrity check on a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectVerifyInput"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectVerification"
  /projects/{id}/quiz:
    post:
      operationId: generateProjectQuiz
      tags: [projects]
      summary: Generate AI interrogation quiz for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectQuiz"
  /projects/{id}/quiz/submit:
    post:
      operationId: submitProjectQuiz
      tags: [projects]
      summary: Submit quiz answers and get trust stamp
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuizSubmission"
      responses:
        "200":
          description: Quiz result and trust stamp
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuizResult"
  /students/{id}/projects:
    get:
      operationId: getStudentProjects
      tags: [projects]
      summary: Get all projects for a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"

  # Certificates
  /students/{id}/certificate:
    get:
      operationId: getCertificate
      tags: [certificates]
      summary: Get competency certificate for a student
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Certificate data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"
        "404":
          description: No certificate yet
  /certificates/{code}:
    get:
      operationId: getCertificateByCode
      tags: [certificates]
      summary: Get certificate by QR code (public)
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Certificate found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Certificate"

  # Dashboard
...
components:
  schemas:
...
    Project:
      type: object
      required: [id, studentId, studentName, title, description, technologies, status, trustStamp, createdAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        title:
          type: string
        description:
          type: string
        technologies:
          type: array
          items:
            type: string
        githubUrl:
          type: ["string", "null"]
        demoUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, analyzing, verified, rejected]
        trustStamp:
          type: boolean
        aiScore:
          type: ["number", "null"]
        integrityNotes:
          type: ["string", "null"]
        createdAt:
          type: string
    ProjectInput:
      type: object
      required: [studentId, title, description, technologies]
      properties:
        studentId:
          type: integer
        title:
          type: string
        description:
          type: string
        technologies:
          type: array
          items:
            type: string
        githubUrl:
          type: string
        demoUrl:
          type: string
    ProjectVerifyInput:
      type: object
      required: [codeSnippet]
      properties:
        codeSnippet:
          type: string
        projectDescription:
          type: string
    ProjectVerification:
      type: object
      required: [projectId, passed, aiScore, summary, originality]
      properties:
        projectId:
          type: integer
        passed:
          type: boolean
        aiScore:
          type: number
        summary:
          type: string
        originality:
          type: number
        integrityNotes:
          type: string
    ProjectQuiz:
      type: object
      required: [projectId, questions]
      properties:
        projectId:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/QuizQuestion"
    QuizQuestion:
      type: object
      required: [id, question, options]
      properties:
        id:
          type: integer
        question:
          type: string
        options:
          type: array
          items:
            type: string
    QuizSubmission:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            $ref: "#/components/schemas/QuizAnswer"
    QuizAnswer:
      type: object
      required: [questionId, answer]
      properties:
        questionId:
          type: integer
        answer:
          type: string
    QuizResult:
      type: object
      required: [projectId, passed, score, feedback, trustStampGranted]
      properties:
        projectId:
          type: integer
        passed:
          type: boolean
        score:
          type: number
        feedback:
          type: string
        trustStampGranted:
          type: boolean
    Certificate:
      type: object
      required: [id, studentId, studentName, major, overallScore, qrCode, skills, projects, issuedAt]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        major:
          type: string
        overallScore:
          type: number
        readinessLevel:
          type: string
        qrCode:
          type: string
        skills:
          type: array
          items:
            $ref: "#/components/schemas/CertifiedSkill"
        projects:
          type: array
          items:
            $ref: "#/components/schemas/CertifiedProject"
        issuedAt:
          type: string
    CertifiedSkill:
      type: object
      required: [name, level, verified, aiVerified]
      properties:
        name:
          type: string
        level:
          type: integer
        verified:
          type: boolean
        aiVerified:
          type: boolean
    CertifiedProject:
      type: object
      required: [title, technologies, trustStamp]
      properties:
        title:
          type: string
        technologies:
          type: array
          items:
            type: string
        trustStamp:
          type: boolean
...