# lib/api-zod/src/generated/types/certifiedProject.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * النبض المهني - Professional Pulse API
 * OpenAPI spec version: 0.1.0
 */

export interface CertifiedProject {
  title: string;
  technologies: string[];
  trustStamp: boolean;
}