# lib/api-zod/src/generated/types/certifiedSkill.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * النبض المهني - Professional Pulse API
 * OpenAPI spec version: 0.1.0
 */

export interface CertifiedSkill {
  name: string;
  level: number;
  verified: boolean;
  aiVerified: boolean;
}