# lib/api-zod/src/generated/types/projectVerification.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * النبض المهني - Professional Pulse API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectVerification {
  projectId: number;
  passed: boolean;
  aiScore: number;
  summary: string;
  originality: number;
  integrityNotes?: string;
}